Move PublicKey to grouper.entities.user
There doesn't seem to be a point in keeping this in a separate
module right now, since only User uses it.  We can always split it
out later.
