--- grouper/entities/public_key.py
@@ -1,5 +1,0 @@
-from typing import NamedTuple
-
-PublicKey = NamedTuple(
-    "PublicKey", [("public_key", str), ("fingerprint", str), ("fingerprint_sha256", str)]
-)
